pyproject.toml
[project]
name = "160grab"
version = "0.1.0"
description = "健康160全自动挂号脚本 - Playwright headless browser automation"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.50.0",
    "httpx>=0.28.1",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-playwright>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grab"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests",
    "live: tests that talk to the real 91160 site",
]

# Ruff configuration
[tool.ruff]
line-length = 88

extend-exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
]
ignore = ["E501", "W505", "F403"]

fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
